Add PlayerCollision and EnemyCollision classes Modify EnemyMain and PlayerHealth
PlayerCollision is responsible for checking player
collisions with enemies and pickups. At the moment,
it checks whether the collision was with an enemy,
and then applies damage to the player.

EnemyCollision will eventually be responsible for
for detecting collisions with player attacks, and
then applying damage to the enemy. At the moment,
it's not really used.

PlayerHealth modified to return maxHealth.

EnemyMain returns damage values.