Modify PlayerMain, PlayerAge, and GameProperties
Added new Delegate into GameProperties, ReturnInt.

PlayerMain has 2 new methods, SetPlayerAge and
GetPlayerAge. SetPlayerAge sets the _currentAge
varable by casting the input as enum AgeState.
This method is subscribed to PlayerAge delegate,
AgeChanged.

GetPlayerAge returns the _currentAge as a casted
integer. This method is subscribed to PlayerAge
delegate, AgeRequested.

PlayerAge has a new method, SetPlayerAge, and 2
new delegates, AgeRequested and AgeChanged.
The current remaining time is checked, and the
appropriate AgeState is sent as a casted integer.
This is done by invoking AgeChanged.