Add method to update Hourglass ui
Add new class UIHourGlass. This updates
the Hourglass UI element as the amount of
time changes.

It has two methods, InvertHourGlass and
UpdateHourGlass. InvertHourGlass sets the
the enum HourGlassInverted variable,
_isHourGlassInverted. _isHourGlassInverted
is used to select whether a slider element
is increasing or decreasing.

UpdateHourGlass updates the slider values
of the sand elements in the Hourglass.
Depending on the value of
_isHourGlassInverted, a slider's value may
be increased or decreased.

InvertHourGlass is subscribed to
PlayerAge delegate, HourGlassFlipped.
UpdateHourGlass is subscribed to PlayerAge
delegate, HourGlassUpdated.

PlayerAge method FlipHourGlass modified to
invoke new delegates HourGlassFlipped,
and HourGlassUpdated.

Countdown IEnumerator modified to
invoke HourGlassUpdated after decrementing
_currentTime.