Fix issue where time is not sent on health death
Sound effects string in UIManager were not
correctly spelt. Changed from Pickup to
PickUp

Countdown Coroutine in PlayerAge was not being
stopped on health death. Added extra delegate
in PlayerHealth that invokes a delegate that
calls a new method in PlayerAge that stops
Countdown coroutine and sends time survived to
UIManager.